+++ src/components/GCChart/index.ts
@@ -1,0 +1,3 @@
+import GCChart from './GCChart';
+
+export default GCChart;