+++ src/components/GCStatusIndicator/index.ts
@@ -1,0 +1,3 @@
+import GCStatusIndicator from './GCStatusIndicator';
+
+export default GCStatusIndicator;