+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Export all components from this repo.
+
 ## 2024-01-30 - 0.4.0
 
 - Added raw (unformatted) and CSV/JSON output to results table.